Add visual indicator to status bar when a fatal failure occurs.
When a fatal failure occurs, a error indicator is added to the status bar, flashes red once, and then turns back to transparent background.

A modal dialog will open when the indicator is clicked, reporting the error, and providing a way to go to the error in the document.

The status bar indicator will go away as soon as Reporter.prototype.checkFatal returns false (no more fatal error).